SHIR-109 # comment split viewer.dart
